+++ src/main/java/com/coffee_shop/coffeeshop/service/coupon/applyservice/CouponApplyService.java
@@ -1,4 +1,4 @@
-package com.coffee_shop.coffeeshop.service.coupon;
+package com.coffee_shop.coffeeshop.service.coupon.applyservice;
 
 import java.time.LocalDateTime;
